+++ perun/fuzz/__init__.py
@@ -1,0 +1,7 @@
+"""Collective package for fuzz performance testing
+
+Contains actual fuzzing rules for different types of workloads,
+the fuzzing loop and strategies/heuristics for enqueueing newly discovered
+workloads for further fuzzing.
+"""
+__author__ = 'Tomas Fiedor'